Ability to categorize networks
